backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.3
python-multipart==0.0.6

# Computer Vision dependencies for disease prediction
opencv-python==4.8.1.78
Pillow==10.1.0
tensorflow-cpu==2.16.1
scikit-image==0.22.0

# MongoDB dependencies
pymongo==4.6.1
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0
